refactor(migration-dialogs): ManualGuideDialog를 별도 파일로 분리
대형 파일 정리 4번째(Phase D). migration_dialogs.py(1697줄)는 완전히
독립적인 ManualGuideDialog(수동 SQL 가이드 다이얼로그)와, 상태 공유가
깊은 단일 거대 클래스 MigrationAnalyzerDialog(1245줄)로 구성됨.
SQLEditorDialog와 동일한 이유로 MigrationAnalyzerDialog 자체는 이번
패스에서 쪼개지 않고, 완전히 독립적인 ManualGuideDialog만 추출.

- migration_manual_guide_dialog.py (신규): ManualGuideDialog
- migration_dialogs.py: 1697 -> 1433줄 (MigrationAnalyzerDialog +
  MigrationWizard + 헬퍼함수/상수 전체 재수입)

검증: 관련 pytest(test_migration_worker.py, test_oneclick_rust_core_gate.py)
37 passed, 전체 pytest 1809 passed / 1 known-flaky 제외. py_compile 통과.
ManualGuideDialog/MigrationAnalyzerDialog 오프스크린 오픈 확인. 앱 스모크 성공.
